refactor: decompose lib/plugins/discord.ts into focused modules
The Discord plugin (1264 lines) was a god file mixing config, rate limiting,
agent pool management, inbound handlers, slash commands, and outbound routing.

Decomposed into lib/plugins/discord/:
- core.ts        — config types, loaders, utilities, registerSlashCommands
- rate-limit.ts  — in-memory + SQLite rate limiter with reconfigure support
- agent-pool.ts  — multi-bot client pool (DISCORD_BOT_TOKEN_* env vars)
- dm-warming.ts  — DM channel pre-warming logic
- inbound.ts     — guild MessageCreate, DM handler, 📋 reaction handler
- slash-commands.ts — /memory command, autocomplete, HITL button interactions
- outbound.ts    — reply routing, webhook posting, HITL renderer registration

lib/plugins/discord.ts is now a 181-line plugin shell that wires submodules.
No behavior changes — purely organizational. All methods preserved.
